Allow filtering by card states
Defaults to active states relevant to the delivery manager.

Filters returned objects rather than using the API, because not all
API endpoints allow passing in with_state option. e.g. if there is
a `filter` option, the API returns an error.

Ignores state if Icebox is selected, as it's the same as state == 'Unscheduled'